cli/pyproject.toml
[project]
name = "notion-agent-mcp"
version = "0.1.0"
description = "MCP server for managing Notion AI Agent instructions"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.10.0",
    "pyyaml",
]

[project.scripts]
notion-agent-mcp = "mcp_server:main"
